+++ pmd-modelica/src/main/java/net/sourceforge/pmd/lang/modelica/ModelicaTokenManager.java
@@ -6,13 +6,20 @@ package net.sourceforge.pmd.lang.modelica;
 
 import java.io.Reader;
 
+import net.sourceforge.pmd.annotation.InternalApi;
 import net.sourceforge.pmd.lang.TokenManager;
 import net.sourceforge.pmd.lang.ast.AbstractTokenManager;
 import net.sourceforge.pmd.lang.ast.SimpleCharStream;
 import net.sourceforge.pmd.lang.modelica.ast.ModelicaParserTokenManager;
 
 
+/**
+ * @deprecated This is internal API
+ */
+@Deprecated
+@InternalApi
 public class ModelicaTokenManager implements TokenManager {
+
     private final ModelicaParserTokenManager modelicaParserTokenManager;
 
     public ModelicaTokenManager(final Reader source) {